First part of re-factoring to use subclasses of CollisionObject representing different shapes like Circle, Rect2, Segment2 etc instead of handling it through he elaborate and confusing and way too fancy Command pattern.
Command pattern is retained but will only be used to determine what to do when events like collision, pick are done. Not to handle collision checking itself. Implications of this is that hierarcical structures no longer handle intersect(Circle), intersect(Rect2) etc. This is just done in simple CollisionObjects. The functions are retained in base class but calling them on a hierarchy will cause assertion failure.