At the moment we can construct roadmap, but we can't pick a source and goal location on map and then find a path through roadmap between that.
This will be the next task after this checkin. 

This checkin takes care of defining dijkstra and A* algorithms. These will later be merged into one because they are so similar in implementation. The interface for Graphs and graph algorithms was changed. 

Before we assumed nothing about the interface of a graph node. Instead each graph algorithm would be provided with functions which would define the way to access edges, calculate distances etc. This was somewhat inspired by boost graph library.

However this seemed to overcomplicate things. A lot of functions had to be passed eventually, and really the flexibility it provided wasn't really needed. So we changed to a more OO oriented solution were it is assumed that nodes have a minimum interface, defined by GraphNode. 

Dependency to CGAL and boost was removed as these libraries were used seldom, and Trapezoidal map was the only one using them and it has fallen out of favor of a probabilistic roadmap.